Added comments and debugged a few things, still in need of testing, Added gitignore to not push .env file and a README to help others set up a working environment
