+++ .gitignore
@@ -1,0 +1,5 @@
+# Python Environment Files
+Python
+
+#ENV File contains API keys
+.env
