+++ README.md
@@ -1,2 +1,35 @@
 # EbayItemFinderPybot
 Ebay bot created in Python to find items
+
+## Installation
+
+Python 3.11.1
+pip install ebaysdk
+pip install pywin32
+pip install python-dotenv
+
+Add a .env file to your file structure
+Inside add api_key = 'YOUR EBAY API KEY'
+
+## Usage
+
+## Finds Currently Listed Items and returns dictionary
+EbayFinderBot
+
+## TODO
+## Finds Previously Sold Items and returns dictionary
+EbayFindSoldItems
+
+## Connects to Visiual Studio Application for Data transfer
+Pipe_Communicator
+
+## TODO
+## Database and transferring Data retrieved to Database
+
+## Contributing
+Uhh Stuff... If you want to contribute lets talk pull requests are welcome
+
+## Fluff
+Creating a bot to also use with the Visual Studio WPF application for Ebay
+Items Finding the best prices to buy and sell. Will atempt future features
+as this project expands.
